Add information from Jenkins CI env to Job json
This includes the run date, the name of the CI job (ci_name), the build
number (ci_id), the URL for the job on the Jenkins dashboard (ci_url)
and the Jenkins job status (status).

Adds pytz package dep since this seems to get a good Py 2.7/3+
compatible way of managing timezones.