Add schemas module to load JSON schema file
This loads the schemas/squash.json verbatim for full the full SQUASH job
json schema.

I've also included functions that dynamically create schemas for the
measurements and packages. It uses the fact that the squash job
schema is modularly structured so that I can extract a subschema from
the 'definitions' property.